% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{core_location}
\alias{core_location}
\title{Location dash-core-components component}
\usage{
core_location(id = NULL, pathname = NULL, refresh = TRUE)
}
\arguments{
\item{id}{(required)}

\item{pathname}{}

\item{refresh}{}
}
