% Auto-generated: do not edit by hand
\name{coreLink}
\alias{coreLink}
\title{Link component}
\description{

}
\usage{
coreLink(..., style=NULL, refresh=NULL, children=NULL, className=NULL, href=NULL, id=NULL)
}
\arguments{
\item{style}{}

\item{refresh}{}

\item{children}{}

\item{className}{}

\item{href}{}

\item{id}{}
}
    