% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{coreLocation}
\alias{coreLocation}
\title{Location dash-core-components component}
\usage{
coreLocation(id = NULL, pathname = NULL, refresh = TRUE)
}
\arguments{
\item{id}{(required)}

\item{pathname}{}

\item{refresh}{}
}
