% Auto-generated: do not edit by hand
\name{coreMarkdown}
\alias{coreMarkdown}
\title{Markdown component}
\description{
A component that renders Markdown text as specified by the CommonMark spec.
}
\usage{
coreMarkdown(..., className=NULL, children=NULL, id=NULL, containerProps=NULL)
}
\arguments{
\item{className}{Class name of the container element}

\item{children}{A markdown string (or array of strings) that adhreres to the CommonMark spec}

\item{id}{}

\item{containerProps}{An object containing custom element props to put on the container
element such as id or style}
}
    