% Auto-generated: do not edit by hand
\name{coreTab}
\alias{coreTab}
\title{Tab component}
\description{

}
\usage{
coreTab(..., className=NULL, style=NULL, disabled_className=NULL, value=NULL, id=NULL, disabled=NULL, selected_className=NULL, selected_style=NULL, label=NULL, disabled_style=list(), children=NULL)
}
\arguments{
\item{className}{Appends a class to the Tab component.}

\item{style}{Overrides the default (inline) styles for the Tab component.}

\item{disabled_className}{Appends a class to the Tab component when it is disabled.}

\item{value}{Value for determining which Tab is currently selected}

\item{id}{The ID of this component, used to identify dash components
in callbacks. The ID needs to be unique across all of the
components in an app.}

\item{disabled}{Determines if tab is disabled or not - defaults to false}

\item{selected_className}{Appends a class to the Tab component when it is selected.}

\item{selected_style}{Overrides the default (inline) styles for the Tab component when it is selected.}

\item{label}{The tab's label}

\item{disabled_style}{Overrides the default (inline) styles when disabled}

\item{children}{The content of the tab - will only be displayed if this tab is selected}
}
    