% Auto-generated: do not edit by hand
\docType{package}
\name{dashCoreComponents-package}
\alias{dashCoreComponents}
\title{Core Interactive UI Components for Dash}
\description{
Dash ships with supercharged components for
interactive user interfaces. A core set of components,
written and maintained by the Dash team, is available in
the dashCoreComponents package. The source for this package
is on GitHub: plotly/dash-core-components.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/plotly/dash-core-components}
  \item Report bugs at \url{https://github.com/plotly/dash-core-components/issues}
}
}
\author{
\strong{Maintainer}: Chris Parmer <chris@plot.ly>
}
