% Auto-generated: do not edit by hand
\name{dccRadioItems}

\alias{dccRadioItems}

\title{RadioItems component}

\description{
RadioItems is a component that encapsulates several radio item inputs. The values and labels of the RadioItems is specified in the `options` property and the seleced item is specified with the `value` property. Each radio item is rendered as an input with a surrounding label.
}

\usage{
dccRadioItems(id=NULL, options=NULL, value=NULL, style=NULL, className=NULL, inputStyle=NULL, inputClassName=NULL, labelStyle=NULL, labelClassName=NULL, loading_state=NULL, ...)
}

\arguments{
\item{id}{}

\item{options}{An array of options}

\item{value}{The currently selected value}

\item{style}{The style of the container (div)}

\item{className}{The class of the container (div)}

\item{inputStyle}{The style of the <input> radio element}

\item{inputClassName}{The class of the <input> radio element}

\item{labelStyle}{The style of the <label> that wraps the radio input
 and the option's label}

\item{labelClassName}{The class of the <label> that wraps the radio input
 and the option's label}

\item{loading_state}{Object that holds the loading state object coming from dash-renderer}
}
