% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{link}
\alias{link}
\title{Link dash-core-components component}
\usage{
link(..., href = NULL, refresh = FALSE, className = NULL, style = NULL,
  id = NULL)
}
\arguments{
\item{...}{children of the component.}

\item{href}{}

\item{refresh}{}

\item{className}{}

\item{style}{}

\item{id}{}
}
