% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{coreInterval}
\alias{coreInterval}
\title{Interval component}
\usage{
coreInterval(id = NULL, interval = 1000, disabled = NULL,
  n_intervals = 0, max_intervals = -1, fireEvent = NULL,
  setProps = NULL, dashEvents = NULL)
}
\arguments{
\item{id}{}

\item{interval}{This component will fire an event every `interval` milliseconds with the event name `setInterval`}

\item{disabled}{If True, the interval will no longer trigger an event.}

\item{n_intervals}{Number of times the interval has passed}

\item{max_intervals}{Number of times the interval will be fired. If -1, then the interval has no limit (the default) and if 0 then the interval stops running.}

\item{fireEvent}{Dash assigned callback}

\item{setProps}{Dash assigned callback}

\item{dashEvents}{}
}
\description{
A component that repeatedly fires an event ("interval") with a fixed time delay between each event. Interval is good for triggering a component on a recurring basis. The time delay is set with the property "interval" in milliseconds.
}
