% Auto-generated: do not edit by hand
\name{coreLink}

\alias{coreLink}

\title{Link component}

\description{

}

\usage{
coreLink(children=NULL, href=NULL, refresh=NULL, className=NULL, style=NULL, id=NULL, loading_state=NULL, ...)
}

\arguments{
\item{children}{}

\item{href}{}

\item{refresh}{}

\item{className}{}

\item{style}{}

\item{id}{}

\item{loading_state}{Object that holds the loading state object coming from dash-renderer}
}
