% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{coreLink}
\alias{coreLink}
\title{Link component}
\usage{
coreLink(..., href = NULL, refresh = FALSE, className = NULL,
  style = NULL, id = NULL, children = NULL)
}
\arguments{
\item{...}{}

\item{href}{}

\item{refresh}{}

\item{className}{}

\item{style}{}

\item{id}{}

\item{children}{}
}
