% Auto-generated: do not edit by hand
\name{coreLink}

\alias{coreLink}

\title{Link component}

\description{

}

\usage{
coreLink(style=NULL, refresh=NULL, className=NULL, href=NULL, id=NULL, ...)
}

\arguments{
\item{style}{}

\item{refresh}{}

\item{className}{}

\item{href}{}

\item{id}{}

\item{...}{wildcards of the form: `data-*` or `aria-*`}
}
