% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{coreLink}
\alias{coreLink}
\title{Link dash-core-components component}
\usage{
coreLink(..., href = NULL, refresh = FALSE, className = NULL,
  style = NULL, id = NULL, children = NULL)
}
\arguments{
\item{...}{(implicit) children of this component, which are combined with the more explicit `children` (a list of components)}

\item{href}{}

\item{refresh}{}

\item{className}{}

\item{style}{}

\item{id}{}

\item{children}{}
}
