% Auto-generated: do not edit by hand
\name{coreLocation}

\alias{coreLocation}

\title{Location component}

\description{
Update and track the current window.location object through the window.history state. Use in conjunction with the `dash_core_components.Link` component to make apps with multiple pages.
}

\usage{
coreLocation(id=NULL, pathname=NULL, search=NULL, hash=NULL, href=NULL, refresh=NULL, ...)
}

\arguments{
\item{id}{}

\item{pathname}{pathname in window.location - e.g., "/my/full/pathname"}

\item{search}{search in window.location - e.g., "?myargument=1"}

\item{hash}{hash in window.location - e.g., "#myhash"}

\item{href}{href in window.location - e.g., "/my/full/pathname?myargument=1#myhash"}

\item{refresh}{Refresh the page when the location is updated?}
}
