% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{coreLocation}
\alias{coreLocation}
\title{Location component}
\usage{
coreLocation(id = NULL, pathname = NULL, search = NULL, hash = NULL,
  href = NULL, refresh = TRUE)
}
\arguments{
\item{id}{(required)}

\item{pathname}{pathname in window.location - e.g., "/my/full/pathname"}

\item{search}{search in window.location - e.g., "?myargument=1"}

\item{hash}{hash in window.location - e.g., "#myhash"}

\item{href}{href in window.location - e.g., "/my/full/pathname?myargument=1#myhash"}

\item{refresh}{Refresh the page when the location is updated?}
}
\description{
Update and track the current window.location object through the window.history state. Use in conjunction with the `dash_core_components.Link` component to make apps with multiple pages.
}
