% Auto-generated: do not edit by hand
\name{coreMarkdown}

\alias{coreMarkdown}

\title{Markdown component}

\description{
A component that renders Markdown text as specified by the CommonMark spec.
}

\usage{
coreMarkdown(id=NULL, className=NULL, containerProps=NULL, dangerously_allow_html=NULL, ...)
}

\arguments{
\item{id}{}

\item{className}{Class name of the container element}

\item{containerProps}{An object containing custom element props to put on the container
element such as id or style}

\item{dangerously_allow_html}{A boolean to control raw HTML escaping.
Setting HTML from code is risky because it's easy to
inadvertently expose your users to a cross-site scripting (XSS)
(https://en.wikipedia.org/wiki/Cross-site_scripting) attack.}
}
