% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{coreMarkdown}
\alias{coreMarkdown}
\title{Markdown dash-core-components component}
\usage{
coreMarkdown(..., id = NULL, className = NULL, containerProps = NULL,
  children = NULL)
}
\arguments{
\item{...}{(implicit) children of this component, which are combined with the more explicit `children` (a list of components)}

\item{id}{}

\item{className}{Class name of the container element}

\item{containerProps}{An object containing custom element props to put on the container element such as id or style}

\item{children}{A markdown string (or array of strings) that adhreres to the CommonMark spec}
}
\description{
A component that renders Markdown text as specified by the CommonMark spec.
}
