% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{coreRadioItems}
\alias{coreRadioItems}
\title{RadioItems dash-core-components component}
\usage{
coreRadioItems(id = NULL, options = list(), value = NULL, style = NULL,
  className = NULL, inputStyle = list(), inputClassName = "",
  labelStyle = list(), labelClassName = "", fireEvent = NULL,
  setProps = NULL, dashEvents = NULL)
}
\arguments{
\item{id}{}

\item{options}{An array of options}

\item{value}{The currently selected value}

\item{style}{The style of the container (div)}

\item{className}{The class of the container (div)}

\item{inputStyle}{The style of the <input> radio element}

\item{inputClassName}{The class of the <input> radio element}

\item{labelStyle}{The style of the <label> that wraps the radio input  and the option's label}

\item{labelClassName}{The class of the <label> that wraps the radio input  and the option's label}

\item{fireEvent}{Dash-assigned callback that gets fired when the radio item gets selected.}

\item{setProps}{Dash-assigned callback that gets fired when the value changes.}

\item{dashEvents}{}
}
\description{
RadioItems is a component that encapsulates several radio item inputs. The values and labels of the RadioItems is specified in the `options` property and the seleced item is specified with the `value` property. Each radio item is rendered as an input with a surrounding label.
}
