% Auto-generated: do not edit by hand
\name{coreRadioItems}
\alias{coreRadioItems}
\title{RadioItems component}
\description{
RadioItems is a component that encapsulates several radio item inputs. The values and labels of the RadioItems is specified in the `options` property and the seleced item is specified with the `value` property. Each radio item is rendered as an input with a surrounding label.
}
\usage{
coreRadioItems(..., labelClassName='', style=NULL, inputClassName='', inputStyle=list(), labelStyle=list(), value=NULL, options=list(), className=NULL, id=NULL)
}
\arguments{
\item{labelClassName}{The class of the <label> that wraps the radio input
 and the option's label}

\item{style}{The style of the container (div)}

\item{inputClassName}{The class of the <input> radio element}

\item{inputStyle}{The style of the <input> radio element}

\item{labelStyle}{The style of the <label> that wraps the radio input
 and the option's label}

\item{value}{The currently selected value}

\item{options}{An array of options}

\item{className}{The class of the container (div)}

\item{id}{}
}
    