% Auto-generated: do not edit by hand
\name{coreTab}

\alias{coreTab}

\title{Tab component}

\description{

}

\usage{
coreTab(id=NULL, label=NULL, value=NULL, disabled=NULL, disabled_style=NULL, disabled_className=NULL, className=NULL, selected_className=NULL, style=NULL, selected_style=NULL, ...)
}

\arguments{
\item{id}{The ID of this component, used to identify dash components
in callbacks. The ID needs to be unique across all of the
components in an app.}

\item{label}{The tab's label}

\item{value}{Value for determining which Tab is currently selected}

\item{disabled}{Determines if tab is disabled or not - defaults to false}

\item{disabled_style}{Overrides the default (inline) styles when disabled}

\item{disabled_className}{Appends a class to the Tab component when it is disabled.}

\item{className}{Appends a class to the Tab component.}

\item{selected_className}{Appends a class to the Tab component when it is selected.}

\item{style}{Overrides the default (inline) styles for the Tab component.}

\item{selected_style}{Overrides the default (inline) styles for the Tab component when it is selected.}
}
